feat(bank ledger): ongoing ledger module implementation

+++ apps/backend/internal/database/schema/bank_ledger.sql
@@ -4,7 +4,7 @@ CREATE TABLE bank_ledger (
     date TIMESTAMPTZ NOT NULL,
     bs_date TEXT NOT NULL,
     entry_type TEXT NOT NULL CHECK (entry_type IN ('dr', 'cr')),
-    amount INTEGER NOT NULL CHECK (amount > 0),
+    amount BIGINT NOT NULL CHECK (amount > 0),
     description TEXT,
     payment_id UUID REFERENCES payments(id) ON DELETE SET NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
